receiver/prometheusreceiver: allow cumulative resets when using the adjuster
Prior to this change, when the start time metric adjuster was used all
points used the same start timestamp. Even after a reset, which makes no
sense for a counter which is not supposed to go down.

Instead this change makes it so that when a reset is detected, the the
reset points timestamp is used as the next start time.

Signed-off-by: Ridwan Sharif <ridwanmsharif@google.com>